Spike using beaker-hostgenerator
Previously, hiera acceptance relied on host definitions in the repo,
and used two different methods for installing puppet-agent. For redhat
and debian, we installed the repo-configs in `install_repos_on` and
installed the packages in `install_packages_on`. For other platforms,
the call to `install_repos_on` would install puppet-agent.

Adding a new platform required adding a host definition file, and
sometimes updating the setup steps for install puppet-agent on that
platform.

This commit modifies the Rakefile to use beaker-hostgenerator to
generate the host definition from an environment variable `PLATFORMS`,
and updates the setup step to call beaker's
`install_puppet_agent_dev_repo` in all cases.

We leave `install_utils` intact, because it is used for `package` style
acceptance. We could probably remove support for that as well.